made string constants and moved activity to session folder
