+++ app/src/main/AndroidManifest.xml
@@ -62,7 +62,7 @@
         <activity
             android:name=".UI.MapsActivity"
             android:label="@string/title_activity_maps" />
-        <activity android:name=".UI.sessionInfoActivity" />
+        <activity android:name=".session.SessionInfoActivity" />
         <activity android:name=".util.BaseMenuActivity"></activity>
     </application>
