+++ app/src/main/res/layout/activity_session_info.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.android.ubclaunchpad.driver.UI.sessionInfoActivity">
+    tools:context="com.android.ubclaunchpad.driver.session.SessionInfoActivity">
 
     <TextView
         android:id="@+id/viewSessionName"
