+++ app/src/main/res/values/strings.xml
@@ -15,8 +15,6 @@
     <string name="register_streetAddress">Street Address (Optional)</string>
     <string name="facebook_app_id">646494705501590</string>
 
-    <string name="session_name">SessionName</string>
-
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
     <string name="passenger_bttn_text">PASSENGER</string>
